feat: add DataDirectoryInitializationMode.SHARED_ROOT_NODE to avoid copying
A new enum case has been added, which ignores versioning on the file
system and behaves in the same way as before this feature.
This enum case will be removed before releasing the final version, but
it's ok to disable copying while folder garbage collection is not
implemented yet.
